gitlab-ci: Add basic build configs.
Define builds for gitlab's continuous integration runners so we
some feedback on versions of the code hosted there, particularly
on https://gitlab.xiph.org/xiph/flac.

Build with GNU Autotools, with CMake+ninja and
verify the Makefile.lite build.

This uses the same Debian-based docker.io/library/gcc:9 container
image we use for most of the codec projects. This doesn't specify
an unprivileged user so that we can install packages, but some
of flac's file-creation tests fail when run as a privileged user.
So we create a user account and use `su` to run the tests.

Another way to address this would be to define a custom
container image with the prerequisites installed before switching
to an unprivileged user for the entire build.